飛行記録作成リファクタリング
