+++ src/app/features/create/flightlog/ui/organisms/FlightDate/types/flightDateTypes.ts
@@ -1,0 +1,7 @@
+export type FlightDateTypes = {
+  handleChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  handleBlur: (e: React.FocusEvent<HTMLInputElement>) => void;
+  values: any;
+  errors: any;
+  touched: any;
+};
