+++ src/app/features/create/flightlog/ui/organisms/FlightDate/validation/validateFlightDate.ts
@@ -1,0 +1,7 @@
+export const validateFlightDate = (value: string) => {
+  let error;
+  if (!value) {
+    error = "飛行日を入力してください";
+  }
+  return error;
+};
