+++ src/app/features/create/flightlog/ui/organisms/FlightPurpose/types/flightPurposeTypes.ts
@@ -1,0 +1,6 @@
+export type FlightPurposeProps = {
+  values: {
+    flightPurpose: string[];
+  };
+  setFieldValue: (field: string, value: any) => void;
+};
