+++ src/app/features/create/flightlog/ui/organisms/FlightSummary/types/flightSummaryFieldTypes.ts
@@ -1,0 +1,6 @@
+export type flightSummaryFieldTypes = {
+  handleChange: (e: React.ChangeEvent<any>) => void;
+  handleBlur: (e: React.FocusEvent<any>) => void;
+  values: any;
+  errors: any;
+};
