+++ src/app/features/create/flightlog/ui/organisms/FlightSummary/validation/validateFlightSummary.ts
@@ -1,0 +1,7 @@
+export const validateFlightSummary = (value: string) => {
+  let error;
+  if (!value) {
+    error = "飛行概要を入力してください";
+  }
+  return error;
+};
