+++ src/app/features/create/flightlog/ui/organisms/LandingLocation/types/landingLocationTypes.ts
@@ -1,0 +1,7 @@
+export type landingLocationTypes = {
+  values: any;
+  setFieldValue: (field: string, value: any) => void;
+  handleChange: (e: React.ChangeEvent<any>) => void;
+  handleBlur: (e: React.FocusEvent<any>) => void;
+  error?: string;
+};
