+++ src/app/features/create/flightlog/ui/organisms/LandingLocation/validation/validateLandingLocation.ts
@@ -1,0 +1,7 @@
+export const validateLandingLocation = (value: string) => {
+  let error;
+  if (!value) {
+    error = "着陸地点を入力してください";
+  }
+  return error;
+};
