+++ src/app/features/create/flightlog/ui/organisms/LandingTimeField/types/landingTimeFieldTypes.ts
@@ -1,0 +1,8 @@
+export type landingTimeFieldTypes = {
+  handleChange: any;
+  handleBlur: any;
+  values: any;
+  errors: any;
+  touched: any;
+  setFieldValue: any;
+};
