+++ src/app/features/create/flightlog/ui/organisms/LandingTimeField/validation/validateFlightDate.ts
@@ -1,0 +1,7 @@
+export const validateFlightDate = (value: string) => {
+  let error;
+  if (!value) {
+    error = "Landing Time is required";
+  }
+  return error;
+};
