+++ src/app/features/create/flightlog/ui/organisms/PilotName/types/PilotNameFieldTypes.ts
@@ -1,0 +1,8 @@
+import { GroupDetail } from "../../types/GroupDetail";
+
+export type PilotNameFieldTypes = {
+  setFieldValue: (field: string, value: any) => void;
+  values: any;
+  fetchGroupDetail: boolean;
+  groupDetail: GroupDetail | null;
+};
