+++ src/app/features/create/flightlog/ui/organisms/PilotName/validation/validatePilotName.ts
@@ -1,0 +1,6 @@
+export const validatePilotName = (value: string) => {
+  if (value === "操縦者を選択してください") {
+    return "操縦者を選択してください";
+  }
+  return "";
+}
