+++ src/app/features/create/flightlog/ui/organisms/TakeoffLocation/types/takeoffLocationTypes.ts
@@ -1,0 +1,7 @@
+export type TakeoffLocationFieldYTypes = {
+  values: any;
+  setFieldValue: any;
+  handleChange: any;
+  handleBlur: any;
+  error?: string;
+};
