+++ src/app/features/create/flightlog/ui/organisms/TakeoffLocation/utility/types/geoLocationTypes.ts
@@ -1,0 +1,11 @@
+export type GeoLocation = {
+  coords: {
+    latitude: number;
+    longitude: number;
+  };
+};
+
+export type Location = {
+  latitude: number;
+  longitude: number;
+};
