+++ src/app/features/create/flightlog/ui/organisms/TakeoffLocation/validation/validateTakeoffLocation.ts
@@ -1,0 +1,7 @@
+export const validateTakeoffLocation = (value: string) => {
+  let error;
+  if (!value) {
+    error = "離陸地点を入力してください";
+  }
+  return error;
+};
