+++ src/app/features/create/flightlog/ui/organisms/TakeoffTime/types/takeoffTimeFieldTypes.ts
@@ -1,0 +1,8 @@
+export type takeoffTimeFieldTypes = {
+  handleChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  handleBlur: (e: React.FocusEvent<HTMLInputElement>) => void;
+  values: any;
+  errors: any;
+  touched: any;
+  setFieldValue: (field: string, value: any) => void;
+};
