+++ src/app/features/edit/flightlog/ui/EditFlightLogTemplate.tsx
@@ -24,7 +24,7 @@ import { TakeOffTimeInputField } from "@/app/features/create/flightlog/ui/molecu
 import { LandingTimeInputField } from "@/app/features/create/flightlog/ui/molecules/LandingTimeInputField";
 import { TotalTimeInputField } from "@/app/features/create/flightlog/ui/molecules/TotalTimeInputField";
 import { handleCalcTime } from "@/app/features/create/flightlog/utility/handleCalcTime";
-import { FlightPurpose } from "@/app/features/create/flightlog/ui/organisms/FlightPurpose";
+import { FlightPurpose } from "@/app/features/create/flightlog/ui/organisms/FlightPurpose/FlightPurpose";
 import { SubmitButton } from "@/app/shared/ui/atoms/SubmitButton/SubmitButton";
 import { format } from "path";
 import { fetchFlightLogUpdate } from "../api/fetchFlightLogUpdate";
